test(broker): extract stub
