+++ broker/src/test/java/io/camunda/zeebe/broker/exporter/stream/ExporterPositionsDistributionTest.java
@@ -2,13 +2,8 @@ package io.camunda.zeebe.broker.exporter.stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.camunda.zeebe.broker.system.partitions.PartitionMessagingService;
-import io.camunda.zeebe.util.collection.Tuple;
-import java.nio.ByteBuffer;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.Executor;
-import java.util.function.Consumer;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -66,28 +61,5 @@ public class ExporterPositionsDistributionTest {
 
     // then
     assertThat(partitionMessagingService.consumers).isEmpty();
-  }
-
-  private static final class SimplePartitionMessageService implements PartitionMessagingService {
-
-    public final Map<String, Tuple<Executor, Consumer<ByteBuffer>>> consumers = new HashMap<>();
-
-    @Override
-    public void subscribe(
-        final String subject, final Consumer<ByteBuffer> consumer, final Executor executor) {
-      consumers.put(subject, new Tuple<>(executor, consumer));
-    }
-
-    @Override
-    public void broadcast(final String subject, final ByteBuffer payload) {
-      final var executorConsumerTuple = consumers.get(subject);
-      final var executor = executorConsumerTuple.getLeft();
-      executor.execute(() -> executorConsumerTuple.getRight().accept(payload));
-    }
-
-    @Override
-    public void unsubscribe(final String subject) {
-      consumers.remove(subject);
-    }
   }
 }
